+++ CHANGELOG.md
@@ -15,6 +15,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 - Tests for `Colors`
 - Tests for `TransformUtils`
 - Tests for `ShapeCreator` (Box, BoxLine, Cylinder, Pointer, Sphere, Arrow, Gizmo)
+- Tests for `DirectShapeUtils`
 
 ## [0.1.0] / 2023-10-01
 ### Quaternion